requirements.txt
pandas
streamlit
seaborn
matplotlib.pyplot